refactor: Completes extraction of auth request handlers.
