+++ api/src/requestHandlers/auth/getClientInfo.js
