+++ api/src/requestHandlers/auth/getGoogleAuthConfig.js
